Move inject-content to common
