+++ controllers/apply/form-router-next/index.js
@@ -23,7 +23,7 @@ const appData = require('../../../common/appData');
 const { localify } = require('../../../common/urls');
 const { noStore } = require('../../../common/cached');
 const { requireActiveUserWithCallback } = require('../../../common/authed');
-const { injectCopy } = require('../../../middleware/inject-content');
+const { injectCopy } = require('../../../common/inject-content');
 
 const salesforceService = require('./lib/salesforce');
 const {
