+++ controllers/home/index.js
@@ -2,7 +2,7 @@
 const path = require('path');
 const express = require('express');
 const contentApi = require('../../common/content-api');
-const { injectCopy } = require('../../middleware/inject-content');
+const { injectCopy } = require('../../common/inject-content');
 
 const router = express.Router();
