+++ controllers/user/activate.js
@@ -11,7 +11,7 @@ const {
 const {
     injectCopy,
     injectBreadcrumbs
-} = require('../../middleware/inject-content');
+} = require('../../common/inject-content');
 
 const logger = require('../../common/logger').child({
     service: 'user'
