+++ controllers/user/index.js
@@ -8,7 +8,7 @@ const { Users } = require('../../db/models');
 const { localify, redirectForLocale } = require('../../common/urls');
 const { noStore } = require('../../common/cached');
 const { requireNotStaffAuth } = require('../../common/authed');
-const { injectCopy } = require('../../middleware/inject-content');
+const { injectCopy } = require('../../common/inject-content');
 
 const logger = require('../../common/logger').child({
     service: 'user'
