+++ controllers/user/password.js
@@ -16,7 +16,7 @@ const { requireNoAuth } = require('../../common/authed');
 const {
     injectCopy,
     injectBreadcrumbs
-} = require('../../middleware/inject-content');
+} = require('../../common/inject-content');
 
 const logger = require('../../common/logger').child({
     service: 'user'
