+++ controllers/user/register.js
@@ -12,7 +12,7 @@ const { csrfProtection } = require('../../common/cached');
 const {
     injectCopy,
     injectBreadcrumbs
-} = require('../../middleware/inject-content');
+} = require('../../common/inject-content');
 const {
     requireNoAuth,
     redirectUrlWithFallback
