+++ controllers/user/update-email.js
@@ -7,7 +7,7 @@ const { Users } = require('../../db/models');
 const { redirectForLocale } = require('../../common/urls');
 const { csrfProtection } = require('../../common/cached');
 const { requireUserAuth } = require('../../common/authed');
-const { injectCopy } = require('../../middleware/inject-content');
+const { injectCopy } = require('../../common/inject-content');
 
 const logger = require('../../common/logger').child({ service: 'user' });
